pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "neutrosophic-pls"
version = "1.0.0"
description = "Neutrosophic Partial Least Squares (N-PLS): Uncertainty-aware PLS with Truth/Indeterminacy/Falsity encoding for chemometrics and spectroscopy"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = [
    "partial-least-squares",
    "pls",
    "neutrosophic",
    "chemometrics",
    "spectroscopy",
    "nir",
    "uncertainty-quantification",
    "machine-learning",
    "regression",
    "variable-importance",
]
authors = [
    {name = "Dickson Abdul-Wahab", email = "dabdul-wahab@live.com"},
    {name = "Ebenezer Aquisman Asare", email = "aquisman1989@gmail.com"},
]
maintainers = [
    {name = "Dickson Abdul-Wahab", email = "dabdul-wahab@live.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=2.0",
]
all = [
    "neutrosophic-pls[dev,docs]",
]

[project.scripts]
npls = "neutrosophic_pls.__main__:main"

[project.urls]
Homepage = "https://github.com/abdulwahab.dickson/neutrosophic-pls"
Documentation = "https://github.com/abdulwahab.dickson/neutrosophic-pls#readme"
Repository = "https://github.com/abdulwahab.dickson/neutrosophic-pls.git"
Issues = "https://github.com/abdulwahab.dickson/neutrosophic-pls/issues"

[tool.setuptools]
packages = ["neutrosophic_pls"]

[tool.setuptools.package-data]
neutrosophic_pls = ["fixtures/*.arff"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

